// src/model/sb-widget-type.enum.ts
export enum SbWidgetType {
  KeyFigureBasic = 'KeyFigureBasic',
  KeyFigureAdvanced = 'KeyFigureAdvanced',
  DataTable = 'DataTable',
  LineChart = 'LineChart',
  StockChart = 'StockChart',
  PieChart = 'PieChart',
  Gauge = 'Gauge',
  ColoredIconBox = 'ColoredIconBox',
  AgDataTable = 'AgDataTable',
  PieChartTable = 'PieChartTable',
  MultiSelect = 'MultiSelect',
  BulletChart = 'BulletChart',
  FormWidget = 'FormWidget',
  FileUpload = 'FileUpload'
}